refactor(workflows): move domain tool files into tools/ subfolder
