+++ synthadoc/agents/workflows/tools/contradiction_resolver_tools.py
